refactor(#3038): extract SessionOverrideState from SharedData (S2)
Lift cluster D — the eight session-scoped override/reset fields
(model_overrides, fast_mode_channels, fast_mode_session_reset_pending,
codex_goals_channels, codex_goals_session_reset_pending,
model_session_reset_pending, session_reset_pending, model_picker_pending)
— out of the SharedData giant into shared_state::SessionOverrideState,
following the S1 QueuedPlaceholderState standard (#3301): verbatim move,
group field at the first member's position, type re-exported from the
module root for surface freeze, constructor literals wrapped at the
first-member position with byte-identical member initializers so the
side-effecting bootstrap evaluation order is unchanged. run_bot's body
is byte-identical to origin/main (verified by brace-balanced extraction
diff); the #3016 hot files have zero diff.

The 41 direct access sites (config.rs 29 / control.rs 5 / fast_mode.rs 1
/ goals.rs 1 / model_picker_interaction.rs 5 — exactly the design's
substitution map) re-route through shared.overrides.<original name>.
Member visibility is re-spelled pub(super) -> pub(in crate::services)
(semantically identical from the new module; documented on the struct).

Design-gap correction (tmux_watcher S1 precedent): the mechanical
substitution alone re-inflates commands/config.rs (its own frozen giant,
baseline 1054) past the ratchet via rustfmt chain splits, and raising a
baseline is forbidden. So the cluster-D analogue of S1's method
co-location applies: the twelve session-override bookkeeping free fns
(reset-entry codec, reset-pending probes/clears, enablement probes,
sync_session_reset_pending) move verbatim to shared_state.rs and are
re-exported from config.rs, keeping every super::config::* importer and
unqualified call site frozen. One documented visibility widening
(fast_mode_reset_entry_matches_provider, module-private ->
pub(in crate::services::discord)); the settings-coupled update_channel_*
writers stay in config.rs.

Accounting: discord/mod.rs giant baseline 4987 -> 4980 (-7, ratchet
down-lock); commands/config.rs 1054 -> 954 (-100, de-gianted: removed
from both giant_file_registry lists, baseline win-locked at 954);
shared_state.rs at 488 prod LoC (< 1000). Inventory docs regenerated,
change-surfaces.md updated (ghost freeze entry removed),
check_agent_maintenance_docs + audit_maintainability --check clean,
await_holding_lock ratchet flat at 34.

The S2-0 characterization tests pass unchanged after the move (the
equivalence proof), as does the pre-existing
bootstrap_session_reset_pending_excludes_restored_model_overrides pin.
