+++ docs/agent-maintenance/change-surfaces.md
@@ -808,7 +808,6 @@
     iteration-loop helpers; tracked decompose target — see
     `giant-file-registry.md` (owner `automation-pipeline`, deadline
     2026-08-31, #3036)).
-  - `src/services/discord/commands/config.rs` (1054 lines).
   - `src/services/discord/{commands/text_commands.rs, commands/diagnostics.rs,
     discord_config_audit.rs, router/intake_gate.rs, inflight.rs}`
     (all 1000+ production lines).
@@ -1047,7 +1046,7 @@ which excludes `#[cfg(test)] mod` blocks); the freshness gate keeps them in sync
   execution are the canonical scheduled JS routine surfaces. Split focused
   helper modules before growing these files again.
 - `src/services/platform/binary_resolver.rs` (1221).
-- `src/services/discord/mod.rs` (4987; +34 from #3019 added the
+- `src/services/discord/mod.rs` (4980; +34 from #3019 added the
   single-authority `increment_global_active` helper + doc mirroring the
   existing decrement helper — offset by removing 6 inline raw `fetch_add`
   blocks across the relay turn-start sites that now route through it; +12 from
@@ -1062,7 +1061,10 @@ which excludes `#[cfg(test)] mod` blocks); the freshness gate keeps them in sync
   the type for surface freeze; ±0 from #3293 — the +13 closed-retry rewires
   (`mailbox_peek` + `*_with_closed_retry` routing for recovery kickoff and
   intervention enqueue) are offset by queue-exit comment dedup in the same
-  root, no baseline raise),
+  root, no baseline raise; -7 from #3038 S2 lifting cluster D — the eight
+  session-override fields — into `shared_state::SessionOverrideState`, leaving
+  a single `overrides: SessionOverrideState` group field on `SharedData` with
+  the type re-exported for surface freeze),
   `src/services/discord_config_audit.rs` (1273).
 - `src/services/turn_orchestrator.rs` (3089; +3 from #3293 declaring the
   `registry_purge` child module — the non-creating `peek` lookup and the
@@ -1078,7 +1080,10 @@ normal test growth is allowed): `src/services/analytics.rs`,
 `src/services/process.rs`, `src/services/discord/tmux_lifecycle.rs`,
 `src/services/qwen_tmux_wrapper.rs`, `src/services/discord/session_relay_sink.rs`,
 `src/services/tui_turn_state.rs`, `src/services/session_backend.rs`,
-`src/voice/turn_link.rs`.
+`src/voice/turn_link.rs`, `src/services/discord/commands/config.rs`
+(#3038 S2: 1054 -> 954 after the session-override bookkeeping helpers
+moved verbatim to `discord/shared_state.rs` next to
+`SessionOverrideState`; still ratcheted at 954 in the frozen baseline).
 
 Same rule: `bugfix` only without a split issue.
