+++ src/services/discord/commands/fast_mode.rs
@@ -69,6 +69,7 @@ pub(in crate::services::discord) async fn cmd_fast(ctx: Context<'_>) -> Result<(
     let reset_pending = ctx
         .data()
         .shared
+        .overrides
         .session_reset_pending
         .contains(&channel_id);
     let notice = if next_enabled {
