+++ src/services/discord/voice_barge_in.rs
@@ -4815,14 +4815,16 @@ mod tests {
             ),
             bot_connected: std::sync::atomic::AtomicBool::new(false),
             last_turn_at: std::sync::Mutex::new(None),
-            model_overrides: dashmap::DashMap::new(),
-            fast_mode_channels: dashmap::DashSet::new(),
-            fast_mode_session_reset_pending: dashmap::DashSet::new(),
-            codex_goals_channels: dashmap::DashSet::new(),
-            codex_goals_session_reset_pending: dashmap::DashSet::new(),
-            model_session_reset_pending: dashmap::DashSet::new(),
-            session_reset_pending: dashmap::DashSet::new(),
-            model_picker_pending: dashmap::DashMap::new(),
+            overrides: super::super::SessionOverrideState {
+                model_overrides: dashmap::DashMap::new(),
+                fast_mode_channels: dashmap::DashSet::new(),
+                fast_mode_session_reset_pending: dashmap::DashSet::new(),
+                codex_goals_channels: dashmap::DashSet::new(),
+                codex_goals_session_reset_pending: dashmap::DashSet::new(),
+                model_session_reset_pending: dashmap::DashSet::new(),
+                session_reset_pending: dashmap::DashSet::new(),
+                model_picker_pending: dashmap::DashMap::new(),
+            },
             dispatch_role_overrides: dashmap::DashMap::new(),
             last_message_ids: dashmap::DashMap::new(),
             catch_up_retry_pending: dashmap::DashMap::new(),
